Add Postman collection and functional testing documentation
- Add Postman collection with all API endpoints organized by category
- Add workflow test suites: Setup Backtest, Execute Trade, Limit Order Flow, End-of-Day Summary
- Add Local and Cloud environment files for Postman/Newman
- Add functional testing guide with rationale for each test
- Fix Polygon API volume parsing (decimal to long conversion)
- Update docs index with testing section

+++ src/AlpacaMock.Infrastructure/Polygon/PolygonClient.cs
@@ -66,7 +66,7 @@ public class PolygonClient : IDisposable
                         High = agg.H,
                         Low = agg.L,
                         Close = agg.C,
-                        Volume = agg.V,
+                        Volume = (long)agg.V,
                         Vwap = agg.Vw,
                         Transactions = agg.N
                     });
@@ -166,7 +166,7 @@ public class PolygonClient : IDisposable
             High = agg.H,
             Low = agg.L,
             Close = agg.C,
-            Volume = agg.V,
+            Volume = (long)agg.V,
             Vwap = agg.Vw,
             Transactions = agg.N
         };
@@ -217,7 +217,7 @@ public class PolygonAgg
     public decimal C { get; set; }  // Close
 
     [JsonPropertyName("v")]
-    public long V { get; set; }  // Volume
+    public decimal V { get; set; }  // Volume (decimal because Polygon sometimes returns floats)
 
     [JsonPropertyName("vw")]
     public decimal? Vw { get; set; }  // VWAP
